requirements.txt
altgraph==0.17
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dash==1.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
Flask==1.1.2
Flask-Compress==1.4.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.18.2
pandas==1.0.3
pefile==2019.4.18
Pillow==7.1.2
plotly==4.5.4
PyInstaller==3.6
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.23.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
urllib3==1.25.9
Werkzeug==1.0.1
